refactor(sqlite): use int16 for stats score fields
- set sqlc types for stats score and score_evaluated columns to int16
- align SQLite repositories and manual query params

+++ miruzo/internal/adapter/persistence/sqlite/imagelist/repository_engaged.go
@@ -43,7 +43,7 @@ func (repo repository) ListEngaged(
 			ctx,
 			gen.ListImagesEngagedParams{
 				Limit:          int64(spec.Limit),
-				ScoreThreshold: int64(spec.ScoreThreshold),
+				ScoreThreshold: spec.ScoreThreshold,
 			},
 		)
 		if err != nil {
@@ -56,9 +56,9 @@ func (repo repository) ListEngaged(
 	rows, err := repo.queries.ListImagesEngagedAfter(
 		ctx,
 		gen.ListImagesEngagedAfterParams{
-			ScoreEvaluated: int64(cursor),
+			ScoreEvaluated: cursor,
 			Limit:          int64(spec.Limit),
-			ScoreThreshold: int64(spec.ScoreThreshold),
+			ScoreThreshold: spec.ScoreThreshold,
 		},
 	)
 	if err != nil {
